+++ src/components/chat/ChatInput.vue
@@ -6,7 +6,7 @@
 <script setup lang="ts">
 import { defineProps, defineEmits, computed, reactive } from "vue";
 import { router } from "@/router";
-import { adjustTooltipPosition } from '@/helpers/tooltipUtils';
+import Tooltip from "@/components/common/Tooltip.vue";
 import useChat from "@/context/chat";
 import useChatbot from "@/context/chatbot";
 import { getChatContentIds } from "@/helpers/contentUtils";
@@ -68,7 +68,7 @@ const toggleListening = () => {
     if (state.speechTimeout)
         clearTimeout(state.speechTimeout);
 
-    if(stt.isListening.value)
+    if (stt.isListening.value)
         stt.stop();
     else
         stt.listen(sttCallback);
@@ -92,17 +92,19 @@ const toggleTTS = () => {
             <div class="chat-content" v-if="getChatContentIds(chat.messages).length > 0" tabindex="0" role="link"
                 @keydown.prevent.space.enter="() => router.push('/content-catalogue')"
                 @click="() => router.push('/content-catalogue')">
-                <div class="icon chatbox-content-default tooltip" @mouseenter="adjustTooltipPosition($event, false)">
-                    <span class="tooltiptext">{{ $t(l.tooltip_open_catalogue_content) }}</span>
-                </div>
+                <Tooltip :text="$t(l.tooltip_open_catalogue_content)" position="left-bottom" :useMaxContent="true"
+                    :adjustPosition="true">
+                    <div class="icon chatbox-content-default">
+                    </div>
+                </Tooltip>
             </div>
             <div v-if="props.optionsVisible" class="chat-options-button"
                 :class="{ 'chat-options-button-active': props.optionsOpen }" role="button"
                 @keydown.prevent.space.enter="$emit('toggleOptions')" @click="$emit('toggleOptions')" tabindex="0">
-                <div class="icon summary-switch-default tooltip"
-                    @mouseenter="adjustTooltipPosition($event, false, 'left-bottom')">
-                    <span class="tooltiptext">{{ $t(l.tooltip_open_chat_side_panel) }}</span>
-                </div>
+                <Tooltip :text="$t(l.tooltip_open_chat_side_panel)" position="left-bottom" :useMaxContent="true"
+                    :adjustPosition="true">
+                    <div class="icon summary-switch-default"></div>
+                </Tooltip>
             </div>
         </div>
         <div class="chat-text-input">
@@ -119,10 +121,11 @@ const toggleTTS = () => {
                 <font-awesome-icon icon="fa-solid fa-volume-high" v-else />
             </div>
             <div class="chat-send-button" @click="submit">
-                <div class="chat-send-icon icon send-message-default tooltip"
-                    @mouseenter="adjustTooltipPosition($event, false)">
-                    <span class="tooltiptext">{{ $t(l.tooltip_send_message) }}</span>
-                </div>
+                <Tooltip :text="$t(l.tooltip_send_message)" position="bottom" :useMaxContent="true"
+                    :adjustPosition="true">
+                    <div class="chat-send-icon icon send-message-default">
+                    </div>
+                </Tooltip>
             </div>
         </div>
     </div>